requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
black==24.4.2
Brotli==1.1.0
click==8.1.7
colorama==0.4.6
decorator==5.1.1
dj-database-url==2.2.0
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
executing==2.0.1
gunicorn==22.0.0
inflection==0.5.1
ipdb==0.13.13
ipython==8.26.0
jedi==0.19.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
packaging==24.1
parso==0.8.4
pathspec==0.12.1
platformdirs==4.2.2
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pure-eval==0.2.2
pycodestyle==2.12.0
Pygments==2.18.0
PyJWT==2.8.0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.19.0
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
wcwidth==0.2.13
whitenoise==6.7.0
